+++ coreblocks/stages/func_blocks_unifier.py
@@ -3,7 +3,7 @@ from collections.abc import Iterable
 from amaranth import *
 
 from coreblocks.params import GenParams, BlockComponentParams
-from transactron.utils.dependencies import UnifierKey, DependencyManager
+from transactron.lib.dependencies import UnifierKey, DependencyManager
 from transactron import Method, TModule
 from transactron.lib import MethodProduct, Collector, Unifier
