+++ transactron/utils/dependencies.py
@@ -1,17 +1,14 @@
 from collections import defaultdict
 
 from abc import abstractmethod, ABC
-from collections.abc import Callable
-from typing import Any, Generic, TypeVar, TYPE_CHECKING
-
-if TYPE_CHECKING:
-    from transactron import Method
-    from transactron.lib.transformers import Unifier
+from typing import Any, Generic, TypeVar
 
 
 __all__ = [
     "DependencyManager",
     "DependencyKey",
+    "SimpleKey",
+    "ListKey"
 ]
 
 T = TypeVar("T")
@@ -83,32 +80,6 @@ class ListKey(Generic[T], DependencyKey[T, list[T]]):
 
     def combine(self, data: list[T]) -> list[T]:
         return data
-
-
-class UnifierKey(DependencyKey["Method", tuple["Method", dict[str, "Unifier"]]]):
-    """Base class for method unifier dependency keys.
-
-    Method unifier dependency keys are used to collect methods to be called by
-    some part of the core. As multiple modules may wish to be called, a method
-    unifier is used to present a single method interface to the caller, which
-    allows to customize the calling behavior.
-    """
-
-    unifier: Callable[[list["Method"]], "Unifier"]
-
-    def __init_subclass__(cls, unifier: Callable[[list["Method"]], "Unifier"], **kwargs) -> None:
-        cls.unifier = unifier
-        return super().__init_subclass__(**kwargs)
-
-    def combine(self, data: list["Method"]) -> tuple["Method", dict[str, "Unifier"]]:
-        if len(data) == 1:
-            return data[0], {}
-        else:
-            unifiers: dict[str, Unifier] = {}
-            unifier_inst = self.unifier(data)
-            unifiers[self.__class__.__name__ + "_unifier"] = unifier_inst
-            method = unifier_inst.method
-        return method, unifiers
 
 
 class DependencyManager:
